moved tests; added extraction function
