+++ src/views/htmlcontent/test/testResources/mockConfig1.spec.ts
@@ -1,4 +1,4 @@
-import { IResultsConfig } from './../interfaces';
+import { IResultsConfig } from './../../src/js/interfaces';
 
 const config: IResultsConfig = {
     shortcuts: {
